requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
filelock==3.19.1
Flask==3.1.3
fonttools==4.60.2
fsspec==2025.10.0
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.7
lazy-loader==0.5
MarkupSafe==3.0.3
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.13.0.92
packaging==26.0
pillow==11.3.0
pip==26.0.1
polars==1.36.1
polars-runtime-32==1.36.1
psutil==7.2.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
scikit-image==0.24.0
scipy==1.13.1
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
tifffile==2024.8.30
torch==2.8.0
torchvision==0.23.0
typing_extensions==4.15.0
ultralytics==8.4.22
ultralytics-thop==2.0.18
urllib3==2.6.3
waitress==3.0.2
Werkzeug==3.1.6
wheel==0.45.1
zipp==3.23.0
